v0/requirements.txt
python-dotenv
fastapi
requests
python-dotenv
uvicorn
